Refactored AuthForm.
