+++ src/pages/AuthPage.tsx
@@ -3,6 +3,8 @@ import AuthForm from "../components/AuthComponents/AuthForm";
 import Loader from "../components/Reusables/Loader";
 import { useAppSelector } from "../store/hooks";
 import { loadingSelector } from "../store/slices/loadingSlice";
+import { auth } from "../firebase/firebase";
+import UserDetails from "../components/AuthComponents/UserDetails";
 
 const AuthPage = () => {
   const isLoading = useAppSelector(loadingSelector);
@@ -16,19 +18,25 @@ const AuthPage = () => {
         <>
           <section className="mt-24 font-outfit flex flex-col justify-center">
             <AuthForm authMode={authMode} />
-            <p className="text-center text-xl uppercase my-1">Or</p>
-            <a
-              className="text-center text-xl text-blue-700 underline "
-              onClick={() => {
-                {
-                  authMode === "Sign Up"
-                    ? setAuthMode("Log In")
-                    : setAuthMode("Sign Up");
-                }
-              }}
-            >
-              {authMode === "Sign Up" ? "Log In" : "Sign Up"}
-            </a>
+            {!auth.currentUser?.isAnonymous ? (
+              <UserDetails />
+            ) : (
+              <>
+                <p className="text-center text-xl uppercase my-1">Or</p>
+                <a
+                  className="text-center text-xl text-blue-700 underline "
+                  onClick={() => {
+                    {
+                      authMode === "Sign Up"
+                        ? setAuthMode("Log In")
+                        : setAuthMode("Sign Up");
+                    }
+                  }}
+                >
+                  {authMode === "Sign Up" ? "Log In" : "Sign Up"}
+                </a>
+              </>
+            )}
           </section>
         </>
       )}
